Updated SensorMicro
Made SerialComms more modular

+++ SensorMicro/Ultra_Sensor.ino
@@ -1,4 +1,4 @@
-long Ultra_Sensor(int P, int E)// Call for the Ultrasonic Sensors
+int Ultra_Sensor(int P, int E)// Call for the Ultrasonic Sensors
 {
  // pinMode(P, OUTPUT); //sets ping pin
   digitalWrite(P, LOW);// These are the ping sent out
@@ -9,7 +9,7 @@ long Ultra_Sensor(int P, int E)// Call for the Ultrasonic Sensors
 
  // pinMode(E, INPUT); //set echo pin
   long  duration = pulseIn(E, HIGH, 11662); // waits for the time duration in microseconds to reach 6 feet out and back or until ping is heard
-  long mm; // Declaration for the millimeters distance because Arduino IDE doesn't like the next declarations
+  int mm; // Declaration for the millimeters distance because Arduino IDE doesn't like the next declarations
   if ( duration > -1) { // if it detects an obstacle closer than 6 feet
     mm = (duration/2) / 2.91; // converts microseconds into length in millimeters
   }
